requirements.txt
Faker==33.1.0
pandas==2.2.3
PyYAML==6.0.2
